% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_prob.R
\name{cluster_prob}
\alias{cluster_prob}
\title{Confidence for Clustering Results}
\usage{
cluster_prob(data, hdat, nclust, nfake = NULL, cutoff = NULL)
}
\arguments{
\item{data}{A matrix or matrix-like object. The `cluster_prob` function treats the
rows as the observations and the columns as the features.}

\item{hdat}{An object of class 'hclust'}

\item{nclust}{number of clusters}

\item{nfake}{number of fake features}

\item{cutoff}{maximum number of important features}
}
\value{
a list containing the following information:
    \item{select_dat}{A matrix containing the data with only important features}
    \item{probability}{A matrix containing the cluster probabilities}
    \item{global_imp}{global features importances}
    \item{class_imp}{class-wise feature importance}
    \item{hdat}{clustering data from `hclust`}
    \item{nclust}{number of clusters}
}
\description{
Calculates probabilities for cluster membership using random forest
}
\details{
This function uses a random forest model to produce a confidence measure for clustering results.
trains a random forest model on the data to predict the clustering labels.
The model uses the features (columns) from the input data to predict the cluster membership of each observation.
}
\examples{
### Example 1: Simulated Data

## Create Data
dat <- matrix(data = rnorm(20*20), nrow = 20)
dat <- dat / max(dat)   # max value = 1
rownames(dat) <- letters[1:20]
colnames(dat) <- LETTERS[1:20]
dat[1:5, ] = dat[1:5, ] + 1    # first cluster
dat[5:10, ] = dat[5:10, ] - 1  # second cluster

## Clustering
hdat <- hclust(dist(dat), method = "ward.D2")

## cluster_prob
prob_result <- cluster_prob(dat, hdat, nclust = 3, nfake = 10, cutoff = 21)

## heatmap_prob
ht <- heatmap_prob(prob_result, -2, 2)

### Example 2: Car Models
car_dat <- mtcars
car_dat <- scale(car_dat)
hdat <- hclust(dist(car_dat))
prob_result <- cluster_prob(car_dat, hdat, 3, 10, 10)

}
