% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_prob.R
\name{heatmap_prob}
\alias{heatmap_prob}
\title{Draws heat map with cluster probabilities}
\usage{
heatmap_prob(result, low_col = -1, high_col = 1)
}
\arguments{
\item{result}{an object produced by 'cluster_prob'}

\item{low_col}{lower expression bound}

\item{high_col}{upper expression bound}
}
\value{
an object of class `HeatmapList` (see \link[pck=ComplexHeatmap]{HeatmapList-class})
}
\description{
Draws heat map with cluster probabilities
}
\examples{
### Example 1: Simulated Data

## Create Data
dat <- matrix(data = rnorm(20*20), nrow = 20)
dat <- dat / max(dat)   # max value = 1
rownames(dat) <- letters[1:20]
colnames(dat) <- LETTERS[1:20]
dat[1:5, ] = dat[1:5, ] + 1    # first cluster
dat[5:10, ] = dat[5:10, ] - 1  # second cluster

## Clustering
hdat <- hclust(dist(dat), method = "ward.D2")
## cluster_prob
prob_result <- cluster_prob(dat, hdat, nclust = 3, nfake = 10, cutoff = 21)
## heatmap_prob
ht <- heatmap_prob(prob_result, -2, 2)

### Example 2: Car Models
car_dat <- mtcars
car_dat <- scale(car_dat)
hdat <- hclust(dist(car_dat))
prob_result <- cluster_prob(car_dat, hdat, 3, 10, 10)

}
\seealso{
\link[pck=ComplexHeatmap]{HeatmapList-class}
}
